% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_theme.R
\name{custom.colorblind}
\alias{custom.colorblind}
\title{Colorblind-safe grey lattice theme}
\usage{
custom.colorblind()
}
\description{
Custom theme for lattice plots. The function takes no arguments.
Colorblind-safe color scale with 7 colors was adapted from R color brewer, 
see \code{RColorBrewer::brewer.pal(8, "Dark2"))}. Color blind safe colors are
distinguishable for most common types of color blindness (deuterotopia, 
deuteroanomaly), yet still look good for non color blind people.
}
\examples{
library(lattice)
data(mtcars)

xyplot(mpg ~ factor(carb) | gear, mtcars,
  groups = carb, auto.key = list(columns = 3),
  par.settings = custom.colorblind()
)
}
