% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_theme.R
\name{custom.ggplot}
\alias{custom.ggplot}
\title{Custom ggplot2 like theme}
\usage{
custom.ggplot(...)
}
\arguments{
\item{...}{graphical parameters passed down to `ggplot2like()` and `simpleTheme()`}
}
\description{
Custom theme for lattice plots. Graphical parameters are passed down
to `ggplot2like()`, which passes them down to `simpleTheme()`.
}
\examples{
library(lattice)
data(mtcars)

xyplot(mpg ~ factor(carb) | gear, mtcars,
  groups = carb, auto.key = list(columns = 3),
  par.settings = custom.ggplot(),
  panel = function(x, y, ...) {
    panel.grid(h = -1, v = -1, col = "white")
    panel.xyplot(x, y, ...)
    panel.lmline(x, y, ...)
  }
)

}
