% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_repelled_labels.R
\name{get_repelled_labels}
\alias{get_repelled_labels}
\title{Get repelled label bounding boxes}
\usage{
get_repelled_labels(
  x,
  y,
  labels,
  cex,
  point_size,
  box_padding_x,
  box_padding_y,
  point_padding_x,
  point_padding_y,
  ...
)
}
\arguments{
\item{x, y}{(numeric) vectors representing x and y coordinates of points}

\item{labels}{(character) vector of labels}

\item{cex}{(numeric) size of text labels}

\item{point_size}{(numeric) approximate size of points in npc relative cooordinates. Can be small,
is used in conjunction with "point_padding" options}

\item{box_padding_x}{(numeric) extra margin around text bounding boxes}

\item{box_padding_y}{(numeric) extra margin around text bounding boxes}

\item{point_padding_x}{(numeric) extra margin around data points}

\item{point_padding_y}{(numeric) extra margin around data points}

\item{...}{other arguments passed to the function}
}
\description{
The function uses ggrepel to obtaine repelled bounding boxes
}
