% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_arrowbox.R
\name{panel.arrowbox}
\alias{panel.arrowbox}
\title{Panel function to draw box like arrows}
\usage{
panel.arrowbox(x0, x1, y0, y1, direction = NULL, tip = NULL, ...)
}
\arguments{
\item{x0, x1, y0, y1}{(numeric) coordinates of the two points determining the 
rectangle (bottom left, top right).}

\item{direction}{(numeric) direction in which arrow should point, 
1 for right, -1 for left side}

\item{tip}{(numeric) pointedness of the arrow head (in native X units. Defaults to a 30th of x scale)}

\item{...}{other arguments passed to the function}
}
\description{
Custom lattice panel function to draw boxes with arrow head from 2 XY 
coordinates and direction argument. Each argument can be a vector of same legnth
so that multiple arrows can be dran at once.
}
\examples{
library(lattice)

xyplot(1:3 ~ 4:6, col = grey(0.5),
  panel = function(x, y, ...) {
    panel.xyplot(x, y, ...)
    panel.arrowbox(x0 = c(4, 5), y0 = c(2.5, 1), 
      x1 = c(5, 6), y1 = c(3, 1.5), 
      direction = c(1, -1), ...)
  }
)

}
