% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_beeswarm.R
\name{panel.beeswarm}
\alias{panel.beeswarm}
\title{Panel function for beeswarm plots}
\usage{
panel.beeswarm(
  x,
  y,
  groups = NULL,
  subscripts = NULL,
  bin_x = FALSE,
  bin_y = FALSE,
  breaks_x = 10,
  breaks_y = 10,
  spread = 0.1,
  ...
)
}
\arguments{
\item{x, y}{(numeric, factor) X and Y variables passed to \code{panel.xyplot()}}

\item{groups}{Passed to \code{panel.xyplot()}.}

\item{subscripts}{Passed to \code{panel.xyplot()}.}

\item{bin_x}{(Logical) If X variable is to be binned or not (default FALSE).}

\item{bin_y}{(Logical) If Y variable is to be binned or not (default FALSE).}

\item{breaks_x}{(numeric) Number of breaks for x variable if \code{bin_x = TRUE}}

\item{breaks_y}{(numeric) Number of breaks for y variable if \code{bin_y = TRUE}}

\item{spread}{(numeric) Scalar indicating how much to spread the values.}

\item{...}{Further arguments passed to \code{xyplot}}
}
\description{
This panel function works essentially like a stripplot, but
instead of randomly scattering a variable produces a regular, grid-like pattern
of points. Currently only the X variable is transformed. However, X and Y 
variables can be discretized optionally.
}
\examples{

library(lattice)

# simple example
df <- data.frame(
  Y = sample(1:10, 60, replace = TRUE), 
  X = factor(rep(1:3, each = 20))
)

xyplot(Y ~ X, df, groups = X, panel = panel.beeswarm)

# but with continuous Y variable, it doesn't work as expected
df$Y <- rnorm(60)
xyplot(Y ~ X, df, groups = X, panel = panel.beeswarm)

# for this purpose we can bin the Y variable into groups
xyplot(Y ~ X, df, groups = X, 
  panel = function(x, y, ...) {
    panel.xyplot(x, y, col = grey(0.6), ...)
    panel.beeswarm(x, y, bin_y = TRUE, breaks_y = 10, ...)
})

}
\author{
modified after a function from Walmes Zeviani, \email{walmes@ufpr.br},
  \code{panel.beeswarm} from package \code{wzRfun}. Idea based on the package
  \code{beeswarm}.
}
