% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_dumbbell.R
\name{panel.dumbbell}
\alias{panel.dumbbell}
\title{Make a dumbbell chart}
\usage{
panel.dumbbell(
  x,
  y,
  groups = NULL,
  subscripts,
  group.number = NULL,
  col_handle = NULL,
  alpha = 1,
  lty = 1,
  lwd = 1,
  direction = NULL,
  ...
)
}
\arguments{
\item{x, y}{(numeric) variables to be plotted}

\item{groups}{grouping variable passed down from xyplot (does not need to be specified)}

\item{subscripts}{subscripts passed down from xyplot (does not need to be specified)}

\item{group.number}{(numeric) internal variable for grouping, does not need to be specified}

\item{col_handle}{(character, numeric) color of the lollipop handle (defaults to superpose.symbol$col)}

\item{alpha, lty, lwd}{(numeric) graphical parameters}

\item{direction}{(character) axis along which the dumbbell is drawn. Default (NULL) is to guess 
the direction from the input variables}

\item{...}{other arguments passed to the function}
}
\description{
This panel function creates dumbbell charts. Similar to a lollipop chart that shows
a single point supported by a horizontal or vertical line, the dumbbell chart shows the
minimum and maximum of a distribution connected by a line. The direction of the dumbbell
(i.e. vertical or horizontal) can be specified manually, or it is determined from the data
automatically (the default).
}
\examples{
library(lattice)
data(mtcars)

# vertical dumbbells
xyplot(mpg ~ factor(gear), mtcars,
  groups = gear,
  pch = 19, lwd = 2,
  panel = function(x, y, ...) {
    panel.dumbbell(x, y, ...)
  }
)

# horizontal dumbbells
xyplot(factor(gear) ~ mpg, mtcars,
  groups = gear,
  pch = 19, lwd = 4, cex = 4,
  panel = function(x, y, ...) {
    panel.dumbbell(x, y, ...)
  }
)

}
