% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_key.R
\name{panel.key}
\alias{panel.key}
\title{Draw custom keys in lattice plots}
\usage{
panel.key(
  groups = NULL,
  labels = NULL,
  which.panel = 1,
  pch = 1,
  cex = 0.8,
  point.cex = NULL,
  points = TRUE,
  lines = FALSE,
  rectangles = FALSE,
  col = NULL,
  lwd = trellis.par.get()$superpose.line$lwd[1],
  lty = trellis.par.get()$superpose.line$lty[1],
  corner = c(0.1, 0.9),
  ...
)
}
\arguments{
\item{groups}{grouping variable passed down from xyplot (does not need to be specified)}

\item{labels}{(character) list of the labels to draw}

\item{which.panel}{(numeric) the panel(s) where key(s) should be drawn}

\item{points}{(logical) if points should be drawn}

\item{lines}{(logical) if lines should be drawn}

\item{rectangles}{(logical) if rectangles should be drawn}

\item{col, lwd, lty, pch, cex, point.cex}{graphical parameters to draw key labels
and symbols}

\item{corner}{(numeric) vector of length 2 indicating the position of the key,
in Normalised Parent Coordinates (0 to 1)}

\item{...}{other arguments passed to the function}
}
\description{
This custom panel function for lattice plots allows to draw
a key (legend) inside a lattice panel, with more customization
options than the lattice default.
}
\examples{
library(lattice)
data(mtcars)

# Two examples for a custom lattice key
# inside a panel. The first with taking all arguments from the 
# top-level plotting function, the second with custom arguments.
xyplot(mpg ~ 1/wt | factor(vs), mtcars,
  groups = carb, pch = 19, cex = 1,
  panel = function(x, y, ...) {
    panel.xyplot(x, y, ...)
    panel.key(...)
    panel.key(labels = letters[1:5], which.panel = 2, 
      corner = c(0.9, 0.1), col = 1:5, pch = 3, cex = 1)
  }
)

}
