% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_piechart.R
\name{panel.piechart}
\alias{panel.piechart}
\title{Draw pie and ring charts in lattice plots}
\usage{
panel.piechart(
  x,
  groups = NULL,
  subscripts = NULL,
  fun = function(x) sum(x, na.rm = TRUE),
  diameter_inner = 0.1,
  diameter_sector = 0.2,
  clockwise = FALSE,
  start_angle = if (clockwise) 90 else 0,
  col = NULL,
  border = NULL,
  lty = NULL,
  lwd = NULL,
  cex = NULL,
  ...
)
}
\arguments{
\item{x}{(numeric, character) variable to be plotted}

\item{groups}{grouping variable passed down from xyplot (does not need to be specified)}

\item{subscripts}{subscripts passed down from xyplot (does not need to be specified)}

\item{fun}{(function) used to perform optional aggregation over groups (default: sum)}

\item{diameter_inner}{(numeric) diameter of the inner circle of the pie/ring (default: 0.1)}

\item{diameter_sector}{(numeric) diameter of the outer circle of the pie/ring (default: 0.2)}

\item{clockwise}{not impemented yet}

\item{start_angle}{not impemented yet}

\item{col, border, lty, lwd, cex}{(character, numeric) graphical parameters}

\item{...}{other arguments passed to the function}
}
\description{
This custom panel function for lattice plots allows to draw
pie charts (or rings) while still being able to use the typical
lattice way of subsetting data. The function can be used
within xyplot() but only one variable needs to be supplied (x).
Grouping is supported in which the x variable is aggregated (summed up) over
each unique group.
}
\examples{
library(grid)
library(lattice)

data("USMortality")

# A simple example using lattice paneling
xyplot( ~ Rate | Sex, USMortality,
  main = "US mortality rates by sex",
  scales = list(draw = FALSE),
  panel = function(x, ...) {
    panel.piechart(x, ...)
  }
)

# A more advanced example using grouping and
# adjusting graphical parameters. The main variable 
# 'x' is now summed up for each value of 'groups'
xyplot( ~ Rate | Sex, USMortality,
  groups = gsub(" ", "\n", Cause), 
  col = heat.colors(10),
  border = grey(0.3), cex = 0.7,
  main = "US mortality rates by sex",
  scales = list(draw = FALSE),
  panel = function(x, ...) {
    panel.piechart(x, diameter_sector = 0.1, ...)
  }
)
}
