% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_symbols.R
\name{panel.symbols}
\alias{panel.symbols}
\title{Plot a grouping variable encoded by symbols}
\usage{
panel.symbols(x, y, z, groups = NULL, subscripts, pch = NULL, ...)
}
\arguments{
\item{x, y}{(numeric) variables to be plotted}

\item{z}{(numeric, factor, character) variable encoding symbols. In not numeric, 
will be coerced to factorand then numeric if possible}

\item{groups}{grouping variable passed down from xyplot (does not need to be specified)}

\item{subscripts}{subscripts passed down from xyplot (does not need to be specified)}

\item{pch}{not used in this context and set to NULL}

\item{...}{other arguments passed to the function}
}
\description{
This panel function allows to plot one additional grouping variable encoded
by symbols ('pch') instead of just the default grouping by color.
Inspired by `panel.bubbleplot` from package `tactile`.
}
\examples{
library(lattice)
data(mtcars)

# first grouping variable for colors, second for symbols (z)
xyplot(mpg ~ factor(cyl), mtcars,
  groups = gear, z = mtcars$cyl,
  panel = function(x, y, z, ...) {
    panel.symbols(x, y, z, ...)
  }
)

}
