% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{repel_boxes}
\alias{repel_boxes}
\title{Adjust the layout of a list of potentially overlapping boxes.}
\usage{
repel_boxes(
  data_points,
  point_size,
  point_padding_x,
  point_padding_y,
  boxes,
  xlim,
  ylim,
  hjust,
  vjust,
  force_push = 1e-07,
  force_pull = 1e-07,
  max_time = 0.1,
  max_overlaps = 10,
  max_iter = 2000L,
  direction = "both",
  verbose = 0L
)
}
\arguments{
\item{data_points}{A numeric matrix with rows representing points like
\code{rbind(c(x, y), c(x, y), ...)}}

\item{point_size}{A numeric vector representing the sizes of data points.}

\item{point_padding_x}{Padding around each data point on the x axis.}

\item{point_padding_y}{Padding around each data point on the y axis.}

\item{boxes}{A numeric matrix with rows representing boxes like
\code{rbind(c(x1, y1, x2, y2), c(x1, y1, x2, y2), ...)}}

\item{xlim}{A numeric vector representing the limits on the x axis like
\code{c(xmin, xmax)}}

\item{ylim}{A numeric vector representing the limits on the y axis like
\code{c(ymin, ymax)}}

\item{hjust}{Optional horizontal adjustment of labels}

\item{vjust}{Optional vertical adjustment of labels}

\item{force_push}{Magnitude of the push force (defaults to \code{1e-7})}

\item{force_pull}{Magnitude of the pull force (defaults to \code{1e-7})}

\item{max_time}{Maximum number of seconds to try to resolve overlaps
(defaults to 0.1)}

\item{max_overlaps}{Maximum number of overlaps without triggering a warning}

\item{max_iter}{Maximum number of iterations to try to resolve overlaps
(defaults to 2000)}

\item{direction}{Character, one of "x", "y", or "both}

\item{verbose}{Boolean,}
}
\description{
Adjust the layout of a list of potentially overlapping boxes.
}
