% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_theme.R
\name{custom.ggplot}
\alias{custom.ggplot}
\title{Custom ggplot2 like theme}
\usage{
custom.ggplot()
}
\description{
Custom theme for lattice plots. The function takes no arguments.
}
\examples{
library(lattice)
data(mtcars)

# colors are more subtle than default lattice theme
xyplot(mpg ~ factor(cyl) | gear, mtcars,
  groups = cyl, auto.key = list(columns = 3),
  par.settings = custom.ggplot()
)
}
