% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_barplot.R
\name{panel.barplot}
\alias{panel.barplot}
\title{Draw barplot with error bars in lattice plots}
\usage{
panel.barplot(
  x,
  y,
  groups = NULL,
  subscripts = NULL,
  col = NULL,
  ewidth = 0.08,
  beside = FALSE,
  origin = NULL,
  FUN_mean = function(x) mean(x, na.rm = TRUE),
  FUN_errb = function(x) sd(x, na.rm = TRUE),
  ...
)
}
\arguments{
\item{x, y}{(numeric, character) variables to be plotted. The x variable 
is treated as a grouping varibale, i.e. error bars are calculated between groups
of unique x values.}

\item{groups}{grouping variable passed down from xyplot (does not need to be specified)}

\item{subscripts}{subscripts passed down from xyplot (does not need to be specified)}

\item{col}{(character) color (vector) to be used for points and lines. 
The default, NULL, uses colors supplied by the top level function.}

\item{ewidth}{(numeric) width of the error bars and whiskers}

\item{beside}{(logical) draw bars/points next to each other (default: FALSE)}

\item{origin}{(numeric) Y coordinate where bars should originate (default NULL means bottom axis)}

\item{FUN_mean}{the function used to calculate group (x-variable) means}

\item{FUN_errb}{the function used to calculate group (x-variable) errors}

\item{...}{other arguments passed to the function}
}
\description{
This custom panel function for lattice plots allows to draw
barplots with error bars for arbitrary groups of data points.
Error bars will be drawn for groups of identical x values with
optional subsetting by grouping or paneling variables.
}
\examples{
library(lattice)
data(mtcars)

# mean and stdev error bars are drawn for
# common x values
xyplot(mpg ~ factor(cyl), mtcars, 
  lwd = 2, pch = 19, cex = 1.5,
  panel = function(x, y, ...) {
    panel.barplot(x, y, ...)
  }
)

# using the same variable for x and grouping will
# result in typical lattice behavior
xyplot(mpg ~ factor(cyl), mtcars, 
  groups = cyl, lwd = 2,
  panel = function(x, y, ...) {
    panel.barplot(x, y, ...)
  }
)

# we can also use different variables for the x var, grouping,
# and paneling. As a visual control that error bars are drawn 
# for the correct groups we overlay the single data points. 
xyplot(mpg ~ factor(cyl) | factor(vs), mtcars,
  groups = gear, lwd = 2, auto.key = list(columns = 3),
  panel = function(x, y, ...) {
    panel.barplot(x, y, beside = TRUE, ...)
    panel.stripplot(x, y, jitter.data = TRUE,
      horizontal = FALSE, amount = 0.15, ...)
  }
)
}
